Moving simplified blocks and signals in files
